damc: use float for everything instead of double
Using double for EqFilter will use more than 20% CPU by just enabling
eq. With floats, it will only use 1.5%.

+++ damc/damc_audio_processing/CompressorFilter.cpp
@@ -27,7 +27,7 @@ void CompressorFilter::init(size_t numChannel) {
 	perChannelData.resize(numChannel);
 }
 
-void CompressorFilter::reset(double fs) {
+void CompressorFilter::reset(float fs) {
 	this->fs = fs;
 	gainHoldSamples = (uint32_t) (fs / 20);
 	std::fill_n(perChannelData.begin(), numChannel, PerChannelData{});
