+++ damc/damc_audio_processing/ExpanderFilter.h
@@ -9,7 +9,7 @@ class ExpanderFilter : public OscContainer {
 public:
 	ExpanderFilter(OscContainer* parent);
 	void init(size_t numChannel);
-	void reset(double fs);
+	void reset(float fs);
 	void processSamples(float** samples, size_t count);
 
 protected:
@@ -23,7 +23,7 @@ private:
 	std::vector<float> previousLevelDetectorOutput;
 
 	OscVariable<bool> enable;
-	double fs = 48000;
+	float fs = 48000;
 	float alphaR;
 	float alphaA;
 	OscVariable<float> attackTime;
